FIX: build errors of stl and eigen
1. Add face property to indexed_triangle_set
2. a few Eigen type errors

+++ src/libslic3r/Format/STEP.cpp
@@ -283,7 +283,7 @@ bool load_step(const char *path, Model *model, ImportStepProgressFn proFn)
             for (Standard_Integer aNodeIter = 1; aNodeIter <= aTriangulation->NbNodes(); ++aNodeIter) {
                 gp_Pnt aPnt = aTriangulation->Node(aNodeIter);
                 aPnt.Transform(aTrsf);
-                points.emplace_back(std::move(Vec3d(aPnt.X(), aPnt.Y(), aPnt.Z())));
+                points.emplace_back(std::move(Vec3f(aPnt.X(), aPnt.Y(), aPnt.Z())));
             }
             //BBS: copy triangles
             const TopAbs_Orientation anOrientation = anExpSF.Current().Orientation();
